+++ odoo/songs/install/post.py
@@ -66,7 +66,7 @@ def setup_mail_catchall_domain(ctx):
 
 
 @anthem.log
-def setup_reports_info(ctx):
+def setup_mts_reports_info(ctx):
     mts = ctx.env.ref('base.main_company')
     mts_logo = 'local-src/mtsmte_reports/static/src/img/mts_logo.jpg'
     if not mts.report_logo:
@@ -75,8 +75,12 @@ def setup_reports_info(ctx):
             'report_show_address_header': False,
         })
         ctx.log_line('MTS report info updated')
+
+
+@anthem.log
+def setup_mte_reports_info(ctx):
     mte = ctx.env.ref('__setup__.company_mte')
-    mte_logo = 'local-src/mtsmte_reports/static/src/img/mte_logo.jpg'
+    mte_logo = 'local-src/mtsmte_reports/static/src/img/mte_logo.png'
     if not mte.report_logo:
         mte.write({
             'report_logo': load_file_content(mte_logo).read().encode('base64'),
@@ -93,4 +97,5 @@ def main(ctx):
     add_company_to_user(ctx)
     create_incoming_mail_server(ctx)
     setup_mail_catchall_domain(ctx)
-    setup_reports_info(ctx)
+    setup_mts_reports_info(ctx)
+    setup_mte_reports_info(ctx)