+++ odoo/songs/update/update_10_0_15.py
@@ -3,7 +3,8 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html)
 
 import anthem
-from ..common import load_csv, load_file_content
+from ..common import load_csv
+from ..install.post import setup_mte_reports_info
 
 
 @anthem.log
@@ -26,21 +27,9 @@ def update_sale_settings(ctx):
 
 
 @anthem.log
-def change_mte_report_logo(ctx):
-    """Update mte report logo"""
-    mte = ctx.env.ref('__setup__.company_mte')
-    mte_logo = 'local-src/mtsmte_reports/static/src/img/mte_logo_new.png'
-    mte.write({
-        'report_logo': load_file_content(mte_logo).read().encode('base64'),
-        'report_show_address_header': True,
-    })
-    ctx.log_line('MTE report info updated')
-
-
-@anthem.log
 def main(ctx):
     """create categories"""
     create_product_categories(ctx)
     update_sale_settings(ctx)
     """update logo"""
-    change_mte_report_logo(ctx)
+    setup_mte_reports_info(ctx)