+++ HISTORY.rst
@@ -32,6 +32,8 @@ latest (unreleased)
 * In the invoice report, comment is now displayed as html
   and won't be displayed if empty BSMTS-215
 * Changed report logo for the mte company BSMTS-213
+* There is no longer default value for comment in 
+  invoice BSMTS-216
 
 **Bugfixes**